job/requirements.txt
pandas==2.3.3
requests==2.32.5
google-cloud-bigquery==3.38.0
google-cloud-storage==3.4.0
db-dtypes==1.4.3